+++ x/crosschain/keeper/abci.go
@@ -30,7 +30,7 @@ type CheckAndUpdateCctxGasPriceFunc func(
 // The function returns the number of cctxs updated and the gas price increase flags used
 func (k Keeper) IterateAndUpdateCctxGasPrice(
 	ctx sdk.Context,
-	chains []*zetachains.Chain,
+	chains []zetachains.Chain,
 	updateFunc CheckAndUpdateCctxGasPriceFunc,
 ) (int, observertypes.GasPriceIncreaseFlags) {
 	// fetch the gas price increase flags or use default
@@ -45,12 +45,14 @@ func (k Keeper) IterateAndUpdateCctxGasPrice(
 		return 0, gasPriceIncreaseFlags
 	}
 
+	additionalChains := k.GetAuthorityKeeper().GetAdditionalChainList(ctx)
+
 	cctxCount := 0
 
 IterateChains:
 	for _, chain := range chains {
 		// support only external evm chains
-		if zetachains.IsEVMChain(chain.ChainId) && !zetachains.IsZetaChain(chain.ChainId) {
+		if zetachains.IsEVMChain(chain.ChainId, additionalChains) && !zetachains.IsZetaChain(chain.ChainId, additionalChains) {
 			res, err := k.ListPendingCctx(sdk.UnwrapSDKContext(ctx), &types.QueryListPendingCctxRequest{
 				ChainId: chain.ChainId,
 				Limit:   gasPriceIncreaseFlags.MaxPendingCctxs,