+++ x/lightclient/types/message_enable_verification_flags.go
@@ -4,12 +4,14 @@ import (
 	cosmoserrors "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	"github.com/zeta-chain/zetacore/pkg/chains"
 )
 
 const (
 	TypeMsgEnableHeaderVerification = "enable_header_verification"
+
+	// MaxChainIDListLength is the maximum number of chain IDs that can be enabled for header verification
+	// this is a value chosen arbitrarily to prevent abuse
+	MaxChainIDListLength = 200
 )
 
 var _ sdk.Msg = &MsgEnableHeaderVerification{}
@@ -46,21 +48,12 @@ func (msg *MsgEnableHeaderVerification) ValidateBasic() error {
 	if _, err := sdk.AccAddressFromBech32(msg.Creator); err != nil {
 		return cosmoserrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
 	}
-	chainListForHeaderSupport := chains.ChainListForHeaderSupport()
 	if len(msg.ChainIdList) == 0 {
 		return cosmoserrors.Wrapf(sdkerrors.ErrInvalidRequest, "chain id list cannot be empty")
 	}
-	if len(msg.ChainIdList) > len(chainListForHeaderSupport) {
-		return cosmoserrors.Wrapf(sdkerrors.ErrInvalidRequest, "chain id list cannot be greater than supported chains")
-	}
-	for _, chainID := range msg.ChainIdList {
-		if !chains.ChainIDInChainList(chainID, chainListForHeaderSupport) {
-			return cosmoserrors.Wrapf(
-				sdkerrors.ErrInvalidRequest,
-				"invalid chain id header not supported (%d)",
-				chainID,
-			)
-		}
+
+	if len(msg.ChainIdList) > MaxChainIDListLength {
+		return cosmoserrors.Wrapf(sdkerrors.ErrInvalidRequest, "chain id list too long")
 	}
 
 	return nil